+++ app/helpers/application_helper.rb
@@ -3,7 +3,14 @@ module ApplicationHelper
   delegate :advanced_search_path, to: :blacklight_advanced_search_engine
 
   # @return [String]
-  def browse_collections_url
-    'https://dss.lafayette.edu/collections'
+  def browse_collections_path
+    search_catalog_path(f: { has_model_ssim: ['Collection'] })
+  end
+
+  # @param [SolrDocument] document
+  # @return [Array<String>]
+  def extracted_text_highlight_values_for(document)
+    return [] unless document.has_highlight_field? 'extracted_text_tsimv'
+    document.highlight_field('extracted_text_tsimv').reject(&:blank?)
   end
 end