+++ config/environments/test.rb
@@ -39,4 +39,5 @@ Rails.application.configure do
   config.active_support.deprecation = :stderr
 
   config.active_job.queue_adapter = :test
+  config.exceptions_app = ->(env) { ErrorController.action(:show).call(env) }
 end